redeployment issues
applied a temporary fix to the issues with api deployment during terraform apply: the api was not being re-deployed even when resources and other things that the api definition is implicitly dependent on change because changes in these resources were not being capture in the hash of the aws_api_gateway_rest_api body on line 198. A better fix would be to use a hash that took into consideration all resources which the gateway depends on.